Avoid overflow of bound

+++ src/main.c
@@ -86,11 +86,11 @@ int main(int argc, char *argv[]) {
     LLL(B, N);
 
     GS_Info *gs_info = gram_schmidt(B, N);
-    double bound = lambda_1(gs_info->Bs, N);
-    printf("Bound: %.6f\n", bound);
+    long double bound = lambda_1(gs_info->Bs, N);
+    printf("Bound: %.6Lf\n", bound);
 
     double result = schorr_euchner(N, gs_info, bound);
-    printf("%8.2f\n", result);
+    printf("%8.8f\n", result);
 
     writeResultToFile(result);
